ADEN-1535 Remove the server name from pattern URL.
