+++ src/test/java/com/wikia/webdriver/Common/Core/URLBuilder/UrlBuilder.java
@@ -77,4 +77,8 @@ public class UrlBuilder {
 		}
 		return url;
 	}
+
+	public String removeProtocolServerNameFromUrl(String url) {
+		return url.substring(url.indexOf('.') + 1, url.length());
+	}
 }
