+++ src/test/java/com/wikia/webdriver/Common/DataProvider/Mobile/MobileAdsDataProvider.java
@@ -10,7 +10,43 @@ import java.util.Arrays;
 public class MobileAdsDataProvider {
 
 	@DataProvider
-	public static Object[][] articlesWithTopLeaderboard() {
+	public static final Object[][] leaderboardAndPrefooterSlots() {
+		return new Object[][]{
+			{
+				"adtest", "SyntheticTests/Slots/leaderboard+prefooter",
+				"wka.ent/_adtest//article",
+				"googlesyndication.com/simgad/8216620376696319112",
+				"googlesyndication.com/pagead/imgad?id=CICAgKCNj62dEhCsAhj6ASgBMgjBw3U0lR5Thg"
+			}
+		};
+	}
+
+	@DataProvider
+	public static final Object[][] leaderboardAndInContentSlots() {
+		return new Object[][]{
+			{
+				"adtest", "SyntheticTests/Slots/leaderboard+in_content",
+				"wka.ent/_adtest//article",
+				"googlesyndication.com/simgad/8216620376696319112",
+				"googlesyndication.com/pagead/imgad?id=CICAgKCNj62dEhCsAhj6ASgBMgjBw3U0lR5Thg"
+			}
+		};
+	}
+
+	@DataProvider
+	public static final Object[][] allSlots() {
+		return new Object[][]{
+			{
+				"adtest", "SyntheticTests/Slots/leaderboard+in content+prefooter",
+				"wka.ent/_adtest//article",
+				"googlesyndication.com/simgad/8216620376696319112",
+				"googlesyndication.com/pagead/imgad?id=CICAgKCNj62dEhCsAhj6ASgBMgjBw3U0lR5Thg"
+			}
+		};
+	}
+
+	@DataProvider
+	public static final Object[][] articlesWithTopLeaderboard() {
 		return new Object[][] {
 			{"elderscrolls", "Skyrim"},
 			{"dragonvale", "Eggs"},
@@ -34,9 +70,8 @@ public class MobileAdsDataProvider {
 		};
 	}
 
-
 	@DataProvider
-	public static Object[][] dfpParams() {
+	public static final Object[][] dfpParams() {
 		return new Object[][] {
 			{
 				"adtest",
@@ -64,7 +99,4 @@ public class MobileAdsDataProvider {
 			}
 		};
 	}
-
-
-
 }
